requirements.txt
# Fantasy Football Draft Assistant - Dependencies
# Updated with all packages used in the project

# Core CrewAI multi-agent framework
crewai==0.152.0
crewai-tools==0.0.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP and async support  
aiohttp==3.9.1
requests==2.31.0

# Web UI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0

# CLI interface and display
click==8.1.7
rich==13.7.1

# Data handling and validation
pandas==2.1.4
pydantic==2.5.2

# AI/LLM integrations
anthropic==0.21.3
openai>=1.13.3  # Used by CrewAI
litellm==1.74.3  # Used by CrewAI

# CrewAI dependencies
chromadb>=0.5.23
instructor>=1.3.3
json-repair==0.25.2
tiktoken>=0.7.0
tenacity>=8.2.3

# Development and testing
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0

# MCP Protocol (for FantasyPros integration)
mcp==1.0.0
fastmcp==0.4.0

# Web scraping for FantasyPros data (fallback only - official MCP server preferred)
beautifulsoup4==4.12.2

# Official FantasyPros MCP Server Dependencies (Node.js based)
# - Install separately: npm install in temp-fantasypros-mcp/
# - Requires FANTASYPROS_API_KEY environment variable
# - Provides: get_rankings, get_players, get_projections, get_sport_news

# Optional: For future Yahoo integration
# yfpy==13.0.0
# yahoo-oauth==2.0